Revert "[TASK] Raise TYPO3 CMS compatibility to 7.1.99"
Some of the changes which come with TYPO3 CMS 7 are not consistent with
TYPO3 CMS 6.2 LTS. Crystalis will be split into different versions
which cover LTS releases and the current sprint release.

On the analogy of system extensions, the version number of Crystalis
will be synchronized with TYPO3 CMS in future.

Releases: 2.0