[BUGFIX] Apply image setup as copy
For an unknown reason, TYPO3 CMS throws an error if the FILES-object of the Image Content Element is referenced by the Textpic Content Element and – in addition – if the FILES-object has got some stdWrap properties. This behavior only seems to occur, if at least 2 Textpic Content Elements appear on a single page. However, it does work fine, if the FILES-object is copied instead of referenced.

Releases: 7.6